+++ internal/distro/distro.go
@@ -90,7 +90,7 @@ var (
 
 func DiskByLabelDir() string { return diskByLabelDir }
 
-func KernelCmdlinePath() string { return kernelCmdlinePath }
+func KernelCmdlinePath() string { return fromEnv("KERNEL_CMDLINE_PATH", kernelCmdlinePath) }
 func BootIDPath() string        { return bootIDPath }
 func SystemRuntimeConfigDir() string {
 	return fromEnv("SYSTEM_RUNTIME_CONFIG_DIR", systemRuntimeConfigDir)